Refactored PublicSuffix test to apply them using both normalization techniques
This reveals that the IdnMappingNormalizer does not have the same 'bug' as using the Uri technique
