+++ Nager.PublicSuffix.UnitTest/Nager.PublicSuffix.UnitTest.csproj
@@ -57,6 +57,8 @@
     <Compile Include="AsyncAwaitTest.cs" />
     <Compile Include="PublicSuffixTest.cs" />
     <Compile Include="DomainNameTest.cs" />
+    <Compile Include="PublicSuffixTestsWithIdnMappingNormalization.cs" />
+    <Compile Include="PublicSuffixTestsWithUriNormalization.cs" />
     <Compile Include="TldRuleTest.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ParseTest.cs" />
