+++ Nager.PublicSuffix.UnitTest/PublicSuffixTest.cs
@@ -1,25 +1,15 @@
 ﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System.IO;
-using System.Linq;
 
 namespace Nager.PublicSuffix.UnitTest
 {
-    [TestClass]
-    public class PublicSuffixTest
+    public abstract class PublicSuffixTest
     {
-        private DomainParser _domainParser;
+        protected DomainParser _domainParser;
 
         //Run tests as specified here:
         //https://raw.githubusercontent.com/publicsuffix/list/master/tests/test_psl.txt
 
-        [TestInitialize()]
-        public void Initialize()
-        {
-            var domainParser = new DomainParser(new FileTldRuleProvider("effective_tld_names.dat"));
-            this._domainParser = domainParser;
-        }
-
-        private void CheckPublicSuffix(string domain, string expected)
+        protected void CheckPublicSuffix(string domain, string expected)
         {
             Assert.IsNotNull(this._domainParser, "_domainParser is null");
 
@@ -161,21 +151,6 @@ namespace Nager.PublicSuffix.UnitTest
             this.CheckPublicSuffix("blogspot.co.ke", null);
             this.CheckPublicSuffix("web.blogspot.co.ke", "web.blogspot.co.ke");
             this.CheckPublicSuffix("a.b.web.blogspot.co.ke", "web.blogspot.co.ke");
-        }
-
-        [TestMethod]
-        public void UnderscoreCheck()
-        {
-            this.CheckPublicSuffix("_abc.def.ghi.jkl.com", "jkl.com");
-            this.CheckPublicSuffix("abc._def.ghi.jkl.com", "jkl.com");
-            this.CheckPublicSuffix("def._ghi.jkl.com", "jkl.com");
-
-            // These tests demonstrate unwanted behaviour due to a 'bug' in Uri construction
-            // def._ghi.jkl.com is valid but
-            // abc.def._ghi.jkl.com is invalid.
-            // It can't be correct that adding abc. in front of a valid domain makes it invalid
-            this.CheckPublicSuffix("abc.def._ghi.jkl.com", null);
-            this.CheckPublicSuffix("abc.def.ghi._jkl.com", null);
         }
     }
 }
